iter5: work with comments
