+++ internal/entities/agent/entites.go
@@ -1,0 +1,5 @@
+package agent
+
+const (
+	CounterMetricName = "Counter"
+)
