+++ internal/models/storage_models.go
@@ -1,0 +1,6 @@
+package models
+
+type StorageValue struct {
+	Counter int64
+	Gauge   float64
+}
