+++ internal/service/server/metricCollector.go
@@ -1,4 +1,4 @@
-package service
+package server
 
 import (
 	"go-metric-svc/dto"
