+++ internal/storage/storage.go
@@ -3,22 +3,18 @@ package storage
 import (
 	"go-metric-svc/dto"
 	"go-metric-svc/internal/customErrors"
+	"go-metric-svc/internal/models"
 	"go.uber.org/zap"
 	"strconv"
 )
 
-type StorageValue struct {
-	Counter int64
-	Gauge   float64
-}
-
 type MemStorage struct {
-	metricsMap map[string]StorageValue
+	metricsMap map[string]models.StorageValue
 
 	log *zap.SugaredLogger
 }
 
-func NewMemStorage(metricsMap map[string]StorageValue, log *zap.SugaredLogger) *MemStorage {
+func NewMemStorage(metricsMap map[string]models.StorageValue, log *zap.SugaredLogger) *MemStorage {
 	return &MemStorage{
 		metricsMap: metricsMap,
 		log:        log,
@@ -27,16 +23,16 @@ func NewMemStorage(metricsMap map[string]StorageValue, log *zap.SugaredLogger) *
 
 func (m *MemStorage) UpdateValue(metricName string, metricValue float64) {
 	m.log.Info("Update in storage")
-	m.metricsMap[metricName] = StorageValue{Gauge: metricValue}
+	m.metricsMap[metricName] = models.StorageValue{Gauge: metricValue}
 }
 
 func (m *MemStorage) SumValue(metricName string, metricValue int64) {
 	if value, exists := m.metricsMap[metricName]; exists {
-		m.metricsMap[metricName] = StorageValue{
+		m.metricsMap[metricName] = models.StorageValue{
 			Counter: value.Counter + metricValue,
 		}
 	} else {
-		m.metricsMap[metricName] = StorageValue{
+		m.metricsMap[metricName] = models.StorageValue{
 			Counter: metricValue,
 		}
 	}
